sensors/requirements.txt
nats-py==2.3.1
